fix : 금액 타입 변경

See merge request s13-fintech-finance-sub1/S13P21A407!75

+++ pocketc_backend/src/main/java/com/ssafy/pocketc_backend/domain/main/dto/MainResponse.java
@@ -1,7 +1,7 @@
 package com.ssafy.pocketc_backend.domain.main.dto;
 
 public record MainResponse(
-        int savingScore,
-        int totalSpending
+        Long savingScore,
+        Long totalSpending
 ) {}
